Check if new page has redirect from old page
The more natural way to check the move, and the way I initially
investigated, is to verify whether the old page redirects to the new one
directly. However, as best I can tell, there’s no easy way to do this:

- The `info` API only tells us _that_ a page is a redirect, not what it
  redirects too (`"redirect": ""`).
- In the page content itself, many wikis translate the `#REDIRECT`
  keyword, so we’d need to know all the translations in order to do this
  reliably.

We could, of course, combine the two: Check if the page is a redirect,
and if it is, match the content on something like

    .* \[\[\(.*\)]].*

but that involves two web requests, and is generally cumbersome.

Instead, we get all all redirects _to_ the new page (the API has a
method for that – I don’t know why, but they must have their reasons,
and I assume it’s implemented efficiently), and check if the old page is
one of them. At the moment, we don’t at all handle the case of there
being more results than the request limit (I think the keyword is
“continuation”), so it’s at least theoretically possible that we will
mistakenly not patrol a page move because the new page has so many other
redirects to it that we don’t see the old page in them. It is, however,
highly unlikely (in my opinion), and anyways it’s not really a big deal
at all if the bot misses to patrol some page moves: it would only be bad
if it patrolled _too many_ revisions.