Remember already seen changes, don’t check again
We store the timestamp of the latest seen revision in a file and, when
getting the changes, tell the API we’re not interested in changes older
than that one (falling back to “15 minutes ago” if the file isn’t
present).

This not only directly removes load on the API (depending on how
frequently the bot is run, we typically get a lot less than 50 results
now – it might well be feasible to run the bot each minute), but also
removes an extra API request, since we now get the changes only once and
cache the result locally instead of requesting it anew each time.

(Annoyingly, the API gives us a formatted YYYY-MM-DDThh:mm:ssZ
timestamp, but requires an unformatted 14-digit timestamp, so we have to
`tr`im away the formatting characters. Not too hard, fortunately :) )